+++ index.js
@@ -1,20 +1,17 @@
 var RSVP = require('rsvp')
 
-module.exports = function promiseMapSeries (arr, iterator, thisArg) {
-  return RSVP.Promise.all(arr)
-    .then(function (arr) {
-      var results = new Array(arr.length)
-      var index = 0
-      return arr.reduce(function (promise, item) {
-          return promise.then(function () {
-              return iterator.call(thisArg, item, index, arr)
-            })
-            .then(function (result) {
-              results[index++] = result
-            })
-        }, RSVP.resolve())
-        .then(function () {
-          return results
+module.exports = function promiseMapSeries (array, iterator, thisArg) {
+  var results = new Array(array.length)
+  var index = 0
+  return array.reduce(function (promise, item) {
+      return promise.then(function () {
+          return iterator.call(thisArg, item, index, array)
         })
+        .then(function (result) {
+          results[index++] = result
+        })
+    }, RSVP.resolve())
+    .then(function () {
+      return results
     })
 }