+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # master
 
+* Do not take promises in `array`
+
 # 0.1.0
 
 * Initial release